requirements.txt
boto==2.48.0
dj-database-url==0.4.2
Django==1.11.8
django-model-utils==3.0.0
django-storages==1.6.5
django-webpack-loader==0.5.0
djangorestframework==3.7.3
python-decouple==3.1
zappa==0.45.1
